requirements.txt
python-socketio
websocket-client